pep8 changes
